% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{persian_palettes}
\alias{persian_palettes}
\title{Persian Color Palettes}
\format{
A named list of character vectors containing hex color codes
}
\usage{
persian_palettes
}
\description{
A list of color palettes inspired by Persian art and artifacts.
Each palette contains a vector of hex color codes.
}
\seealso{
\code{\link{persian_palette}} to generate palettes and \code{\link{persian_palettes_colorblind_safe}} for colorblind-safe options.
}
\keyword{datasets}
