% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{persian_palettes_colorblind_safe}
\alias{persian_palettes_colorblind_safe}
\title{Colorblind-Safe Persian Color Palettes}
\format{
A named list of character vectors containing hex color codes
}
\usage{
persian_palettes_colorblind_safe
}
\description{
A subset of Persian color palettes that are colorblind-safe.
These palettes have been tested for accessibility, see \code{vignette("check-colors")}.
}
\keyword{datasets}
