% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{persian_palette}
\alias{persian_palette}
\title{A Palette Generator Based on Persian Art}
\usage{
persian_palette(
  name,
  n,
  direction = c(1, -1),
  selection = c("sequential", "evenly")
)
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired. Empty or -1 for all colors.}

\item{direction}{Sets the order of colors in the palette. If 1, the default,
colors are as output in the palette. If -1, the order of colors is reversed.}

\item{selection}{For discrete palettes, either "sequential" (default, selects adjacent colors)
or "evenly" (maximizes distance between selected colors).}
}
\value{
A vector of colours.
}
\description{
This is a collection of color palettes based on artifacts of persian art.
}
