% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{persian_palette}
\alias{persian_palette}
\title{A Palette Generator Based on Persian Art}
\usage{
persian_palette(
  name,
  n,
  type = c("discrete", "continuous"),
  direction = c(1, -1)
)
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours}
}
\value{
A vector of colours.
}
\description{
This is a collection of color palettes based on artifacts of persian art.
}
